# be/api/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  fiber.Map:
    additionalProperties: true
    type: object
...
  http.ConfirmUploadRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      object_key:
        type: string
      pod_id:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - object_key
    - pod_id
    - title
    type: object
  http.CreateVersionRequest:
    properties:
      changelog:
        maxLength: 1000
        type: string
      object_key:
        type: string
    required:
    - object_key
    type: object
...
  http.UpdateMaterialRequest:
    properties:
      description:
        maxLength: 2000
        type: string
      title:
        maxLength: 255
        type: string
    type: object
...
  http.UploadURLRequest:
    properties:
      content_type:
        type: string
      filename:
        type: string
      size:
        description: 100MB max
        maximum: 104857600
        type: integer
    required:
    - content_type
    - filename
    - size
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@ngasihtau.com
    name: NgasihTau API Support
  description: |-
    NgasihTau is a learning platform API that enables teachers to create Knowledge Pods for sharing learning materials with students.
    The platform uses a microservices architecture with Go as the primary backend language.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: NgasihTau API
  version: "1.0"
paths:
...
  /materials/{id}:
    delete:
      consumes:
      - application/json
      description: Soft-delete a material
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Material deleted
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Delete a material
      tags:
      - Materials
    get:
      consumes:
      - application/json
      description: Get a material by ID
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Material details
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Material not found
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Get a material
      tags:
      - Materials
    put:
      consumes:
      - application/json
      description: Update a material's title and description
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Material update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UpdateMaterialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated material
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Update a material
      tags:
      - Materials
...
  /materials/{id}/download:
    get:
      consumes:
      - application/json
      description: Generate a presigned URL for downloading a material
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Download URL
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Material not found
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Get download URL
      tags:
      - Materials
  /materials/{id}/preview:
    get:
      consumes:
      - application/json
      description: Generate a presigned URL for previewing a material
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Preview URL
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/fiber.Map'
        "404":
          description: Material not found
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Get preview URL
      tags:
      - Materials
...
  /materials/{id}/versions:
    get:
      consumes:
      - application/json
      description: Get all versions of a material
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of versions
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid material ID
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Get version history
      tags:
      - Materials
    post:
      consumes:
      - application/json
      description: Upload a new version of an existing material
      parameters:
      - description: Material ID
        format: uuid
        in: path
        name: id
        required: true
        type: string
      - description: Version details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.CreateVersionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created version
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Create a new version
      tags:
      - Materials
  /materials/confirm:
    post:
      consumes:
      - application/json
      description: Confirm a file upload and create a material record
      parameters:
      - description: Upload confirmation details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.ConfirmUploadRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created material
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request or upload error
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Confirm file upload
      tags:
      - Materials
  /materials/upload-url:
    post:
      consumes:
      - application/json
      description: Generate a presigned URL for direct file upload to storage
      parameters:
      - description: Upload request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/http.UploadURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Upload URL and object key
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid request or unsupported file type
          schema:
            $ref: '#/definitions/fiber.Map'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: Get presigned upload URL
      tags:
      - Materials
...
  /pods/{podId}/materials:
    get:
      consumes:
      - application/json
      description: Get a paginated list of materials in a Knowledge Pod
      parameters:
      - description: Pod ID
        format: uuid
        in: path
        name: podId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        maximum: 100
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: List of materials
          schema:
            $ref: '#/definitions/fiber.Map'
        "400":
          description: Invalid pod ID
          schema:
            $ref: '#/definitions/fiber.Map'
      security:
      - BearerAuth: []
      summary: List materials in a pod
      tags:
      - Materials
...
securityDefinitions:
  BearerAuth:
    description: 'JWT Bearer token. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"